chore: split out util functions for clarity
